refactor: module organization
