+++ src/index.js
@@ -1,0 +1,3 @@
+import {default as getVideoId} from './get-video-id';
+
+export default getVideoId;
